refactoring getFaucetBalance
